Work on User

+++ src/main/java/com/stepanwxw/crud/view/RegionMenu.java
@@ -1,9 +1,14 @@
 package main.java.com.stepanwxw.crud.view;
 
 import main.java.com.stepanwxw.crud.model.Region;
+import main.java.com.stepanwxw.crud.model.User;
 import main.java.com.stepanwxw.crud.repository.RegionRepositoryImpl;
+import main.java.com.stepanwxw.crud.repository.UserRepositoryImpl;
 
 import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Scanner;
 
 public class RegionMenu {
@@ -52,10 +57,22 @@ public class RegionMenu {
                     System.out.println("Enter Id for delete: ");
                     try {
                         long id = Long.parseLong(lineInput());
-                        regionRepository.remove(id);
-                        System.out.println("Congratulation. Id " + id + " is delete.");
+                        UserRepositoryImpl userRepository = new UserRepositoryImpl();
+                        List<User> userList = new ArrayList<>(userRepository.getAll());
+                        int ind = 0;
+                        for (User u : userList) {
+                            if (u.getRegion().getId() == id) {
+                                System.out.println("Cannot be deleted, because region id = " + id + " use from User.");
+                                ind = 1;
+                            }
+                        }
+                            if (!(ind == 1)) {regionRepository.remove(id);
+                                System.out.println("Congratulation. Id " + id + " is delete.");
+                            }
                     } catch (NumberFormatException e) {
                         System.out.println("Input number please");
+                    } catch (IOException e) {
+                        throw new RuntimeException(e);
                     }
                     indicator = false;
                     break;
